requirements (1).txt
pyTelegramBotAPI==4.14.1
telethon>=1.28.0
rapidfuzz>=3.9.0
python-dotenv==1.0.0